[Refactor] 필드명 통일 및 가격 단위 변경 (만원 → 원)
## 필드명 통일
- Member.projectCountType → projectCount로 일관성 개선
- ProfileResponse, PartnerProfileResponse, CreateProjectResponse 모두 통일

## 가격 단위 변경 (만원 → 원)
### 타입 변경
- PartnerRecruit.cost: Integer → Long
- ConfirmProjectRequest.totalEstimate: Integer → Long
- ConfirmProjectRequest.RecruitmentRequest.cost: Integer → Long
- CreateProjectResponse.totalEstimate: Integer → Long
- CreateProjectResponse.EstimateResponse.cost: Integer → Long
- CreateProjectResponse.PartnerResponse.suggestedCost: Integer → Long

### 로직 수정
- ProjectService.estimateCostPerRole(): int → long, 프롬프트 "만원" → "원"
- ProjectService.recommendPartners(): avgCost int → long
- ProjectService.parsePartnerRecommendations(): Integer.parseInt → Long.parseLong
- ProjectService.fallbackRecommendPartners(): int cost → long cost
- PayLogService.prepare(): recruit.getCost() * 10000 → recruit.getCost() (변환 제거)

### 이점
- Integer 오버플로우 방지 (큰 금액 처리)
- DB와 실제 결제 금액 일치 (변환 로직 제거)
- 모든 cost/amount/estimation 필드 원 단위 통일

+++ src/main/java/com/example/off/domain/partnerRecruit/PartnerRecruit.java
@@ -37,7 +37,7 @@ public class PartnerRecruit {
     private Integer numberOfPerson;
 
     @Column(nullable = false)
-    private Integer cost;
+    private Long cost;  // 원 단위
 
     @CreationTimestamp
     @Column(nullable = false, updatable = false)
@@ -62,7 +62,7 @@ public class PartnerRecruit {
     @OneToMany(mappedBy = "partnerRecruit")
     private List<PartnerApplication> partnerApplications = new ArrayList<>();
 
-    public PartnerRecruit(Project project, Role role, Integer numberOfPerson, RecruitStatus recruitStatus, int cost) {
+    public PartnerRecruit(Project project, Role role, Integer numberOfPerson, RecruitStatus recruitStatus, Long cost) {
         this.project = project;
         this.role = role;
         this.numberOfPerson = numberOfPerson;
